Fix file structures
